+++ cobib/commands/init.py
@@ -26,6 +26,8 @@ class InitCommand(Command):
         """
         LOGGER.debug('Starting Init command.')
         parser = ArgumentParser(prog="init", description="Init subcommand parser.")
+        parser.add_argument('-g', '--git', action='store_true',
+                            help="initialize git repository")
         parser.add_argument('-f', '--force', action='store_true',
                             help="DEPRECATED! This argument will be removed in v2.6")
 
@@ -36,20 +38,45 @@ class InitCommand(Command):
             return
 
         if largs.force:
-            msg = 'The "force" argument has been deprecated and will be removed in v2.6\n' + \
-                  'If you would like to overwrite your current database, please edit the file ' + \
-                  'manually.'
+            msg = 'The "force" argument has been deprecated and is in the process of being removed.'
             warnings.warn(msg, DeprecationWarning)
             LOGGER.warning(msg)
 
         conf_database = CONFIG.config['DATABASE']
         file = os.path.realpath(os.path.expanduser(conf_database['file']))
-        if os.path.exists(file) and not largs.force:
-            msg = "Database file already exists! Use --force to overwrite."
-            print(msg, file=sys.stderr)
-            LOGGER.error(msg)
-            return
-        LOGGER.debug('Creating path for database file: "%s"', os.path.dirname(file))
-        os.makedirs(os.path.dirname(file), exist_ok=True)
-        LOGGER.debug('Creating empty database file: "%s"', file)
-        open(file, 'w').close()
+        root = os.path.dirname(file)
+
+        file_exists = os.path.exists(file)
+        git_tracked = os.path.exists(os.path.join(root, '.git'))
+
+        if file_exists:
+            if git_tracked:
+                msg = 'Database file already exists and is being tracked by git. ' + \
+                      'There is nothing else to do.'
+                print(msg, file=sys.stderr)
+                LOGGER.info(msg)
+                return
+
+            if not git_tracked and not largs.git:
+                msg = 'Database file already exists! Use --git to start tracking it with git.'
+                print(msg, file=sys.stderr)
+                LOGGER.warning(msg)
+                return
+
+        else:
+            LOGGER.debug('Creating path for database file: "%s"', root)
+            os.makedirs(root, exist_ok=True)
+
+            LOGGER.debug('Creating empty database file: "%s"', file)
+            open(file, 'w').close()
+
+        if largs.git:
+            if not conf_database.getboolean('git'):
+                msg = 'You are about to initialize the git tracking of your database, but this ' + \
+                      'will only have effect if you also enable the DATABASE/git setting in ' + \
+                      'your configuration file!'
+                print(msg, file=sys.stderr)
+                LOGGER.warning(msg)
+            LOGGER.debug('Initializing git repository in "%s"', root)
+            os.system(f'git init {root}')
+            self.git(args=vars(largs), force=True)