+++ test/test_commands.py
@@ -1,14 +1,16 @@
 """Tests for CoBib's commands."""
 # pylint: disable=unused-argument, redefined-outer-name
 
+import json
 import os
 import re
+import subprocess
 import sys
 from datetime import datetime
 from io import StringIO
 from itertools import zip_longest
 from pathlib import Path
-from shutil import copyfile
+from shutil import rmtree
 
 import pytest
 from cobib import commands
@@ -16,11 +18,29 @@ from cobib.config import CONFIG
 from cobib.database import read_database
 
 
+def assert_git_commit_message(command, args):
+    """Assert the last auto-generated git commit message."""
+    # get last commit message
+    proc = subprocess.Popen(['git', '-C', '/tmp/cobib_test', 'show',
+                             '--format=format:%B', '--no-patch', 'HEAD'],
+                            stdout=subprocess.PIPE)
+    message, _ = proc.communicate()
+    # decode it
+    message = message.decode('utf-8').split('\n')
+    # assert subject line
+    assert f'Auto-commit: {command.title()}Command' in message[0]
+    # assert args
+    args = json.dumps(args, indent=2, default=str)
+    for ref, truth in zip(args.split('\n'), message[2:]):
+        assert ref == truth
+
+
 @pytest.fixture
 def setup():
     """Setup."""
     root = os.path.abspath(os.path.dirname(__file__))
     CONFIG.set_config(Path(root + '/../cobib/docs/debug.ini'))
+    CONFIG.validate()
     read_database()
 
 
@@ -34,68 +54,68 @@ def test_set_config(setup):
     assert CONFIG.config['DATABASE']['file'] == './test/example_literature.yaml'
     # change back to default
     CONFIG.set_config()
+    CONFIG.validate()
     assert CONFIG.config['DATABASE']['file'] == \
         os.path.expanduser('~/.local/share/cobib/literature.yaml')
 
 
-def test_init():
-    """Test init command."""
+@pytest.fixture(params=[False, True])
+def init_setup(request):
+    """Setup for InitCommand testing."""
     # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
+    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test/database.yaml\n"
+    if request.param:
+        tmp_config += 'git=True\n'
     with open('/tmp/cobib_test_config.ini', 'w') as file:
         file.write(tmp_config)
+    # ensure configuration is empty
+    CONFIG.config = {}
+    # load config
     CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
+    CONFIG.validate()
+    # yielding the arguments allows re-using them inside of the actual test function
+    yield request.param
+    # clean up file system
+    os.remove('/tmp/cobib_test_config.ini')
+    if request.param:
+        rmtree('/tmp/cobib_test/.git')
+
+
+@pytest.mark.parametrize(['safe'], [
+        [False],
+        [True],
+    ])
+def test_init(init_setup, safe):
+    """Test init command."""
+    git = init_setup
+    if safe:
+        # fill database file
+        with open('/tmp/cobib_test/database.yaml', 'w') as file:
+            file.write('test')
     # store current time
     now = float(datetime.now().timestamp())
-    commands.InitCommand().execute({})
-    # check creation time of temporary database file
-    ctime = os.stat('/tmp/cobib_test_database.yaml').st_ctime
-    # assert these times are close
-    assert ctime - now < 0.1 or now - ctime < 0.1
-    # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
-
-
-def test_init_safe():
-    """Test init aborts when database file exists."""
-    # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
-    with open('/tmp/cobib_test_config.ini', 'w') as file:
-        file.write(tmp_config)
-    CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
-    # fill database file
-    with open('/tmp/cobib_test_database.yaml', 'w') as file:
-        file.write('test')
     # try running init
-    commands.InitCommand().execute({})
-    # check init aborted and database file still contains 'test'
-    with open('/tmp/cobib_test_database.yaml', 'r') as file:
-        assert file.read() == 'test'
+    commands.InitCommand().execute(['--git'] if git else [])
+    if safe:
+        # check database file still contains 'test'
+        with open('/tmp/cobib_test/database.yaml', 'r') as file:
+            assert file.read() == 'test'
+    else:
+        # check creation time of temporary database file
+        ctime = os.stat('/tmp/cobib_test/database.yaml').st_ctime
+        # assert these times are close
+        assert ctime - now < 0.1 or now - ctime < 0.1
+    if git:
+        # check creation time of temporary database git folder
+        ctime = os.stat('/tmp/cobib_test/.git').st_ctime
+        # assert these times are close
+        assert ctime - now < 0.1 or now - ctime < 0.1
+        # and assert that it is indeed a folder
+        assert os.path.isdir('/tmp/cobib_test/.git')
+        # assert the git commit message
+        assert_git_commit_message('init', {'git': True, 'force': False})
     # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
-
-
-# DEPRECATD: to be removed in v2.6
-def test_init_force():
-    """Test init can be forced when database file exists."""
-    # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
-    with open('/tmp/cobib_test_config.ini', 'w') as file:
-        file.write(tmp_config)
-    CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
-    # fill database file
-    with open('/tmp/cobib_test_database.yaml', 'w') as file:
-        file.write('test')
-    # try running init
-    with pytest.deprecated_call():
-        commands.InitCommand().execute(['-f'])
-        # check init was forced and database file was overwritten
-        assert os.stat('/tmp/cobib_test_database.yaml').st_size == 0
-    # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
+    os.remove('/tmp/cobib_test/database.yaml')
 
 
 @pytest.mark.parametrize(['args', 'expected'], [
@@ -166,6 +186,7 @@ def open_setup():
     CONFIG.config = {}
     root = os.path.abspath(os.path.dirname(__file__))
     CONFIG.set_config(Path(root + '/../cobib/docs/debug.ini'))
+    CONFIG.validate()
     # NOTE: normally you would never trigger an Add command before reading the database but in this
     # controlled testing scenario we can be certain that this is fine
     commands.AddCommand().execute(['-b', './test/dummy_multi_file_entry.bib'])
@@ -207,43 +228,50 @@ def test_open(open_setup):
     sys.stdin = original_stdin
 
 
-def test_add():
-    """Test add command."""
-    # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
-    with open('/tmp/cobib_test_config.ini', 'w') as file:
-        file.write(tmp_config)
-    CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
-    # ensure database file exists and is empty
-    open('/tmp/cobib_test_database.yaml', 'w').close()
+@pytest.fixture
+def database_setup(init_setup):
+    """Initialize a database on top of the init_setup fixture."""
+    git = init_setup
+    # initialize database
+    # NOTE: if the InitCommand fails, all tests depending on this will fail, too
+    commands.InitCommand().execute(['--git'] if git else [])
     # freshly read in database to overwrite anything that was read in during setup()
     read_database(fresh=True)
+    # yield the parameter to allow re-use in actual test function
+    yield git
+    # clean up file system
+    os.remove('/tmp/cobib_test/database.yaml')
+
+
+def test_add(database_setup):
+    """Test add command."""
+    git = database_setup
     # add some data
     commands.AddCommand().execute(['-b', './test/example_literature.bib'])
     # compare with reference file
-    with open('/tmp/cobib_test_database.yaml', 'r') as file:
+    with open('/tmp/cobib_test/database.yaml', 'r') as file:
         with open('./test/example_literature.yaml', 'r') as expected:
             for line, truth in zip_longest(file, expected):
                 assert line == truth
-    # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
+    if git:
+        # assert the git commit message
+        with open('./test/example_literature.bib', 'r') as bibtex:
+            assert_git_commit_message('add', {
+                'label': None,
+                'file': None,
+                'arxiv': None,
+                'bibtex': bibtex,
+                'doi': None,
+                'isbn': None,
+                'tags': [],
+            })
 
 
-def test_add_overwrite_label():
+def test_add_overwrite_label(database_setup):
     """Test add command while specifying a label manually.
 
     Regression test against #4.
     """
-    # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
-    with open('/tmp/cobib_test_config.ini', 'w') as file:
-        file.write(tmp_config)
-    CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
-    # ensure database file exists and is empty
-    open('/tmp/cobib_test_database.yaml', 'w').close()
-    # freshly read in database to overwrite anything that was read in during setup()
-    read_database(fresh=True)
     # add some data
     commands.AddCommand().execute(['-b', './test/example_literature.bib'])
     # add potentially duplicate entry
@@ -254,42 +282,37 @@ def test_add_overwrite_label():
         true_lines = expected.readlines()
     with open('./test/example_duplicate_entry.yaml', 'r') as extra:
         true_lines += extra.readlines()
-    with open('/tmp/cobib_test_database.yaml', 'r') as file:
+    with open('/tmp/cobib_test/database.yaml', 'r') as file:
         for line, truth in zip_longest(file, true_lines):
             assert line == truth
-    # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
 
 
 @pytest.mark.parametrize(['labels'], [
-        ['knuthwebsite'],
+        [['knuthwebsite']],
         [['knuthwebsite', 'latexcompanion']],
     ])
-def test_delete(labels):
+def test_delete(database_setup, labels):
     """Test delete command."""
-    # use temporary config
-    tmp_config = "[DATABASE]\nfile=/tmp/cobib_test_database.yaml\n"
-    with open('/tmp/cobib_test_config.ini', 'w') as file:
-        file.write(tmp_config)
-    CONFIG.set_config(Path('/tmp/cobib_test_config.ini'))
-    # copy example database to configured location
-    copyfile(Path('./test/example_literature.yaml'), Path('/tmp/cobib_test_database.yaml'))
+    git = database_setup
+    # NOTE: DeleteCommand depends on AddCommand to work. While this is not so nice for the
+    # unittests, it is the easiest method of testing all git and non-git scenarios.
+    commands.AddCommand().execute(['-b', './test/example_literature.bib'])
     # delete some data
     # NOTE: for testing simplicity we delete the last entry
     commands.DeleteCommand().execute(labels)
-    with open('/tmp/cobib_test_database.yaml', 'r') as file:
+    with open('/tmp/cobib_test/database.yaml', 'r') as file:
         with open('./test/example_literature.yaml', 'r') as expected:
             # NOTE: do NOT use zip_longest to omit last entry (thus, we deleted the last one)
             for line, truth in zip(file, expected):
                 assert line == truth
             with pytest.raises(StopIteration):
                 file.__next__()
-    # clean up file system
-    os.remove('/tmp/cobib_test_database.yaml')
-    os.remove('/tmp/cobib_test_config.ini')
+    if git:
+        # assert the git commit message
+        assert_git_commit_message('delete', {'labels': labels})
 
 
+# TODO: figure out some very crude and basic way of testing this
 def test_edit():
     """Test edit command."""
     pytest.skip("There is currently no meaningful way of testing this.")